Create seperate class for rendering
