+++ app/src/main/java/com/michaeltroger/gruenerpass/model/PdfRenderer.kt
@@ -6,27 +6,24 @@ import android.graphics.Bitmap
 import android.graphics.Canvas
 import android.graphics.Color
 import android.graphics.pdf.PdfRenderer
-import android.net.Uri
 import android.os.ParcelFileDescriptor
 import com.google.zxing.*
 import com.google.zxing.common.BitMatrix
 import com.google.zxing.common.HybridBinarizer
 import com.google.zxing.qrcode.QRCodeReader
 import com.google.zxing.qrcode.decoder.ErrorCorrectionLevel
-import com.tom_roush.pdfbox.pdmodel.PDDocument
-import com.tom_roush.pdfbox.pdmodel.encryption.InvalidPasswordException
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import java.io.File
-import java.io.FileOutputStream
 
-
-const val PDF_FILENAME = "certificate.pdf"
 private const val QR_CODE_SIZE = 400
 private const val MULTIPLIER_PDF_RESOLUTION = 2
 private const val MAX_BITMAP_SIZE = 100 * 1024 * 1024
 
-class Pdf(private val context: Context) {
+class PdfRenderer(private val context: Context) {
+
+    private val file = File(context.filesDir, PDF_FILENAME)
+
     private val activityManager: ActivityManager?
         get() = context.getSystemService(Context.ACTIVITY_SERVICE) as? ActivityManager
 
@@ -36,22 +33,8 @@ class Pdf(private val context: Context) {
     private var bitmapDocument: Bitmap? = null
     private var bitmapQrCode: Bitmap? = null
 
-    private val file = File(context.filesDir, PDF_FILENAME)
-
     fun getQrBitmap() = bitmapQrCode
     fun getPdfBitmap() = bitmapDocument
-
-    suspend fun doesFileExist(): Boolean = withContext(Dispatchers.IO) {
-        file.exists()
-    }
-
-    suspend fun deleteFile() = withContext(Dispatchers.IO) {
-        if (doesFileExist()) {
-            file.delete()
-        }
-        bitmapDocument = null
-        bitmapQrCode = null
-    }
 
     /**
      * @return true if successful
@@ -70,7 +53,7 @@ class Pdf(private val context: Context) {
                 renderer?.close()
                 fileDescriptor?.close()
             } catch (ignore: Exception) {}
-            deleteFile()
+            //deleteFile() TODO
             return@withContext false
         }
 
@@ -104,7 +87,7 @@ class Pdf(private val context: Context) {
         } catch (ignore: Exception) {}
 
         if (bitmapDocument!!.byteCount > MAX_BITMAP_SIZE) {
-            deleteFile()
+            //deleteFile() TODO
             return@withContext false
         }
 
@@ -148,59 +131,5 @@ class Pdf(private val context: Context) {
         bitmapQrCode = null
         bitmapQrCode = Bitmap.createBitmap(w, h, Bitmap.Config.RGB_565)
         bitmapQrCode!!.setPixels(pixels, 0, QR_CODE_SIZE, 0, 0, w, h)
-    }
-
-    suspend fun isPdfPasswordProtected(uri: Uri): Boolean = withContext(Dispatchers.IO) {
-        try {
-            context.contentResolver.openInputStream(uri)!!.use {
-                try {
-                    return@withContext PDDocument.load(it).checkIfPasswordProtectedAndClose()
-                } catch (exception: InvalidPasswordException) {
-                    return@withContext true
-                }
-            }
-        } catch (exception: Exception) {
-            return@withContext false
-        }
-    }
-    /**
-     * @return true if successful
-     */
-    suspend fun copyPdfToCache(uri: Uri): Boolean = withContext(Dispatchers.IO) {
-        try {
-            context.contentResolver.openInputStream(uri)!!.use {
-                deleteFile() // clear old file first if it exists
-                it.copyTo(FileOutputStream(file))
-                return@withContext true
-            }
-        } catch (exception: Exception) {
-            return@withContext false
-        }
-    }
-
-    /**
-     * @return true if successful
-     */
-    suspend fun decryptAndCopyPdfToCache(uri: Uri, password: String): Boolean = withContext(Dispatchers.IO) {
-        try {
-            context.contentResolver.openInputStream(uri)!!.use {
-                with(PDDocument.load(it, password)) {
-                    deleteFile() // clear old file first if it exists
-                    removePasswordCopyAndClose()
-                }
-                return@withContext true
-            }
-        } catch (exception: Exception) {
-            return@withContext false
-        }
-    }
-
-    private fun PDDocument.removePasswordCopyAndClose() = use {
-        isAllSecurityToBeRemoved = true
-        save(FileOutputStream(file))
-    }
-
-    private fun PDDocument.checkIfPasswordProtectedAndClose(): Boolean = use {
-        return isEncrypted
     }
 }
